% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CERMBlidarpostgis.R
\docType{package}
\name{CERMBlidarpostgis}
\alias{CERMBlidarpostgis}
\alias{CERMBlidarpostgis-package}
\title{CERMBlidarpostgis: Spatial database functions for LIDAR data}
\description{
This package extends the CERMBlidar package with functions for
storing and retrieving rasters of vegetation density to/from a
PostgreSQL / PostGIS database.
}
