% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_setting_functions.R
\name{check_postgis}
\alias{check_postgis}
\title{Check for a local PostgreSQL/PostGIS installation}
\usage{
check_postgis(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{TRUE}, warning messages about missing components are
not printed to the console. The default is \code{FALSE}.}
}
\value{
\code{TRUE} if PostgreSQL/PostGIS is found;
  \code{FALSE} otherwise.
}
\description{
Checks that PostgreSQL version 11 and the PostGIS extension are installed on
the local system and available from the command line.
}
