% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_setting_functions.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Checks for a local PostgreSQL/PostGIS installation and sets runtime variables}
\usage{
connect_to_database(dbname, host = "localhost", port = 5432,
  username = "postgres", password = NULL,
  pgpath = "c:/Program Files/PostgreSQL/11/bin")
}
\arguments{
\item{quiet}{If \code{TRUE}, warning messages about missing components are
not printed to the console. The default is \code{FALSE}.}
}
\value{
\code{TRUE} if PostgreSQL/PostGIS is found;
  \code{FALSE} otherwise.
}
\description{
This function first checks that PostgreSQL version 11 and the PostGIS
extension are installed on the local system and available from the command
line. It then creates a database connection pool to provide for simple
queries from R.
}
