% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_check_childless_metadata}
\alias{db_check_childless_metadata}
\title{Check for any records in metadata tables with no associated point count rasters}
\usage{
db_check_childless_metadata(dbsettings, schema = NULL)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{schema}{A character vector giving the names of one or more schemas
(e.g. "mgazone56") to check. If \code{NULL} (default), all schemas are
checked.}
}
\value{
A data frame. If childless meta-data records were found, the schema,
  integer id and filename will be listed for each. If no such records were
  found the data frame will have zero rows and columns.
}
\description{
Each record in a \code{point_counts} table, which contains rasters of point
counts for vertical strata, is linked to a parent record in the
\code{las_metadata} table within the same schema (e.g. 'mgazone56') via the
relation \code{point_counts.meta_id = las_metadata.id}. The database will not
allow a point counts record without an associated metadata record, but it
will allow a metadata record without any point counts records. Usually, this
will indicate that some problem occurred during import. This function checks
for any such childless metadata records.
}
