% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnect_functions.R
\name{db_connect_postgis}
\alias{db_connect_postgis}
\title{Connect to a PostGIS database holding data derived from LIDAR images}
\usage{
db_connect_postgis(dbname, host = "localhost", port = 5432,
  username = "postgres", password = NULL,
  pgpath = "c:/Program Files/PostgreSQL/11/bin")
}
\arguments{
\item{dbname}{Database name.}

\item{host}{Host name. Defaults to \code{'localhost'}.}

\item{port}{Port number to use for database commands. Defaults to \code{5432}
which is the default port number for PostgreSQL.}

\item{username}{User name for this session. Defaults to \code{'postgres'}.}

\item{password}{Database password for the specified user. If \code{NULL}
(default) this is taken from the local system environment variable
\code{PGPASSWORD}.}

\item{pgpath}{The path to the \code{bin} directory of a locally accessible
installation of PostgreSQL / PostGIS. Defaults to
\code{'c:/Program Files/PostgreSQL/11/bin'}.}
}
\value{
A named list of database settings for use with other \code{db_xxx}
  functions.
}
\description{
This function first checks that PostgreSQL version 11 and the PostGIS
extension are installed on the local system and available from the command
line. It then creates a database connection pool to provide for simple
queries from R and returns this and other database parameters in a named list
to be used with \code{db_xxx} functions.
}
\seealso{
\code{\link{db_create_postgis}}
}
