% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnect_functions.R
\name{db_disconnect_postgis}
\alias{db_disconnect_postgis}
\title{Close an open connection to a PostGIS LiDAR database}
\usage{
db_disconnect_postgis(dbsettings)
}
\arguments{
\item{DBsettings}{A named list of database settings as returned by
\code{db_create_postgis} and \code{db_connect_postgis}.}
}
\description{
This is a convenience function, equivalent to calling
\code{pool::poolClose(DBsettings$POOL)} directly.
}
