% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_export_stratum_counts}
\alias{db_export_stratum_counts}
\title{Export raster point count data for a selected tile.}
\usage{
db_export_stratum_counts(dbsettings, epsg, meta_id, outpath)
}
\description{
Export raster point count data for a selected tile.
}
