% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_import_las}
\alias{db_import_las}
\title{Import a raster of point counts and its metadata into the database}
\usage{
db_import_las(dbsettings, las.path, dem.path = NULL,
  epsg.reproject = NULL, counts.tablename = "rasters.tmp_load",
  metadata.tablename = "vectors.las_metadata")
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{db_connect_postgis}}.}

\item{las.path}{The path and filename of the LAS source file. This can be
an uncompressed (\code{.las}) or compressed \code{.laz} or \code{.zip}
file.}

\item{dem.path}{The path and filename of a DEM raster file to use for
normalizing point heights. This can be an uncompressed (\code{.asc} or
\code{.tif}) file, or a compressed \code{.zip} file. If \code{NULL}
(default), then point heights will be normalized via triangulation of
the LAS ground points using the \code{\link[lidR]{tin}} algorithm
with \code{\link[lidR]{lasnormalize}}.}

\item{epsg.reproject}{If an integer EPSG code is provided, the point cloud
will be reprojected (if necessary) prior to deriving point counts and other
data. If \code{NULL} (default), no reprojection is done.}

\item{counts.tablename}{The name of the table in which to store the raster of
point counts within vertical layers. Should be in the form
\code{'schema.table'}. Defaults to \code{'rasters.tmp_load'}.}

\item{metadata.tablename}{The name of the table in which to store the summary
data and bounding polygon for the LAS image. Should be in the form
\code{'schema.table'}. Defaults to \code{'vectors.las_metadata'}.}
}
\description{
Import a raster of point counts and its metadata into the database
}
