% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_lasfile_imported}
\alias{db_lasfile_imported}
\title{Check if one or more LAS files have been imported into the database}
\usage{
db_lasfile_imported(dbsettings, filenames)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{filenames}{A character vector of one or more file names. These can be
full paths or file names with or without extensions. File extensions
(e.g. \code{.las; .LAZ; .zip}) and are ignored for comparison purposes, as
is case.}
}
\value{
A logical vector with the input file names as element names and
  \code{TRUE} values indicating files that have already been imported.
}
\description{
When a LAS image is imported, the name of the file from which it was read is
recorded in the \code{lidar.las_metadata} table of the database. This
function queries that table to see which, if any, of the names provided
in the \code{filenames} argument are present in that table.
}
\seealso{
\code{\link{db_connect_postgis}} \code{\link{db_create_postgis}}
}
