% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_load_tile_metadata}
\alias{db_load_tile_metadata}
\title{Import LAS tile metadata into the database}
\usage{
db_load_tile_metadata(dbsettings, las, filename, tablename)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{connect_to_database}}.}

\item{las}{A LAS object.}

\item{filename}{Path or filename from which the LAS object was read.}

\item{tablename}{Name of the raster table in the form \code{'schema.tablename'}.}
}
\description{
Import LAS tile metadata into the database
}
