% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{get_point_counts}
\alias{get_point_counts}
\title{Get counts for point classes}
\usage{
get_point_counts(las)
}
\arguments{
\item{las}{A LAS object.}
}
\value{
A named list of the count of points in each of the following
classes: ground, veg, building, water, other.
}
\description{
Gets the number of points in classes: ground, veg, building, water, and
other.
}
