% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{get_raster_bounds}
\alias{get_raster_bounds}
\title{Get the bounding envelope of a raster as a polygon}
\usage{
get_raster_bounds(r)
}
\arguments{
\item{r}{The raster.}
}
\value{
Raster bounds as an \code{sf::polygon} object.
}
\description{
This returns the bounds of a raster as an \code{sf} polygon object,
suitable for writing to PostGIS.
}
