% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{pg_load_raster}
\alias{pg_load_raster}
\title{Import a raster into the database}
\usage{
pg_load_raster(r, epsg, dbname, schema, tbl, replace = FALSE,
  tilew = 50, username = "postgres")
}
\description{
Import a raster into the database
}
