% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_export_point_counts}
\alias{db_export_point_counts}
\title{Export raster point count data that intersects with one or more features}
\usage{
db_export_point_counts(dbsettings, geom, outpath, time.interval = NULL,
  default.epsg = NULL)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{geom}{The feature(s) to intersect. Can also be an \code{Extent} object
or a 2x2 matrix where the first row is X min,max and the second row is
Y min,max.}

\item{outpath}{A path and filename for the output GeoTIFF file.}

\item{time.interval}{If provided, only export data for rasters where the
scan time (capture_start and capture_end fields) overlap the given interval.
Can be provided as either a vector of one or more years as four-digit integers,
or a \code{lubridate::interval} object.}

\item{default.epsg}{EPSG code to assume for the query feature(s). Ignored if
the features are \code{sf} or \code{raster::Extent} objects with a coordinate
reference system defined.}
}
\value{
A raster stack linked to the exported GeoTIFF file.
}
\description{
This function takes a set of features (points, lines, polygons or an extent
in the form of a matrix or raster Extent object), identifies the raster tiles
of point count data that the features intersect with, and exports these to a
GeoTIFF file.
}
\note{
Presently, this function takes raster data from the non-unioned
point counts table where each record corresponds to an imported LAS file.
The data are then unioned before export to allow for overlapping LAS tile
edges. The original intention was to have a second raster point counts
table of pre-unioned data, but this has proved difficult to implement
efficiently and reliably.
}
\seealso{
\code{\link{db_connect_postgis}} \code{\link{db_create_postgis}}
}
