% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_import_las}
\alias{db_import_las}
\title{Import a raster of point counts and its metadata into the database}
\usage{
db_import_las(dbsettings, las.path, dem.path = NULL)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{db_connect_postgis}}.}

\item{las.path}{The path and filename of the LAS source file. This can be
an uncompressed (\code{.las}) or compressed \code{.laz} or \code{.zip}
file.}

\item{dem.path}{The path and filename of a DEM raster file to use for
normalizing point heights. This can be an uncompressed (\code{.asc} or
\code{.tif}) file, or a compressed \code{.zip} file. If \code{NULL}
(default), then point heights will be normalized via triangulation of
the LAS ground points using the \code{\link[lidR]{tin}} algorithm
with \code{\link[lidR]{lasnormalize}}.}
}
\description{
First, the LAS tile is read from disk and the function checks that it is
located in one of the map zones supported by the database (each zone's
data is stored in a separate schema). Next, point heights are normalized,
any overlap between flight lines is removed, and the counts of vegetation,
ground and water points are rasterized as vertical layers defined in
\code{CERMBlidar::CERMBstrata}. Finally, the point count data are imported
into the database along with meta-data for tile extent, point density, etc.
}
