% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_import_las}
\alias{db_import_las}
\title{Import a raster of point counts and its metadata into the database}
\usage{
db_import_las(dbsettings, las.path, epsg.code = 3308)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{las.path}{The path and filename of the LAS source file. This can be
an uncompressed (\code{.las}) or compressed \code{.laz} or \code{.zip}
file.}

\item{epsg.code}{The EPSG code for the coordinate reference system to apply
to data. The point cloud will be re-projected as necessary, prior to
deriving point counts and other data. The default EPSG code is 3308
(New South Wales Lambert projection / GDA94).}
}
\description{
Import a raster of point counts and its metadata into the database
}
\seealso{
\code{\link{db_connect_postgis}} \code{\link{db_create_postgis}}
}
