% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_import_las}
\alias{db_import_las}
\title{Import a raster of point counts and its metadata into the database}
\usage{
db_import_las(dbsettings, las.path, dem.path = NULL,
  counts.tablename = "rasters.tmp_load",
  metadata.tablename = "vectors.las_metadata", epsg.code = 3308)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{connect_to_database}}.}

\item{las.path}{The path and filename of the LAS source file. This can be
an uncompressed (\code{.las}) or compressed \code{.laz} or \code{.zip}
file.}

\item{dem.path}{The path and filename of a DEM raster file to use for
normalizing point heights. This can be an uncompressed (\code{.asc} or
\code{.tif}) file, or a compressed \code{.zip} file. If \code{NULL}
(default), then point heights will be normalized via triangulation of
the LAS ground points using the \code{\link[lidR]{tin}} algorithm
with \code{\link[lidR]{lasnormalize}}.}

\item{counts.tablename}{The name of the table in which to store the raster of
point counts within vertical layers. Should be in the form
\code{'schema.table'}. Defaults to \code{'rasters.tmp_load'}.}

\item{metadata.tablename}{The name of the table in which to store the summary
data and bounding polygon for the LAS image. Should be in the form
\code{'schema.table'}. Defaults to \code{'vectors.las_metadata'}.}

\item{epsg.code}{The EPSG code for the coordinate reference system to apply
to data. The point cloud will be re-projected as necessary, prior to
deriving point counts and other data. The default EPSG code is 3308
(New South Wales Lambert projection / GDA94).}
}
\description{
Import a raster of point counts and its metadata into the database
}
