% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_import_las}
\alias{db_import_las}
\title{Import a raster of point counts and its metadata into the database}
\usage{
db_import_las(dbsettings, las.paths, dem.paths = NULL,
  check.overlap = TRUE, union.rasters = FALSE, union.batch = 1)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{db_connect_postgis}}.}

\item{las.paths}{The paths and filenames of the LAS source files. This can
be an uncompressed (\code{.las}) or compressed \code{.laz} or \code{.zip}
file.}

\item{dem.paths}{The paths and filenames of DEM (digital elevation model)
rasters corresponding to the LAS source files. If a vector of paths is
provided it must be the same length as \code{las.paths}. For each element
that specifies the path to a raster file (e.g. GeoTIFF) or zipped raster
file, the DEM data will read and used to normalize the point heights for
the corresponding LAS file. Where an element is \code{NA} or
an empty string, point heights for the corresponding LAS file will be
normalized using Delaunay triangulation. If this argument is set to
\code{NULL} (default), triangulation will be used for all LAS files.}

\item{check.overlap}{If \code{TRUE} (default), images will be checked for
overlap between flight lines using the function
\code{remove_flightline_overlap} in the \code{CERMBlidar} package. If an
image fails the overlap check (i.e. it appears to have overlap and this
cannot be removed) it will be skipped. Setting this argument to
\code{FALSE} disables overlap checking. This is useful for images with
unusual configurations of flight lines that cause a false positive overlap
check.}

\item{union.rasters}{If \code{TRUE}, raster point counts derived from the
LAS images will be loaded into both the \code{point_counts} table as a
single record and the \code{point_counts_union} table where data for
spatially adjacent, overlapping rasters are merged. If \code{FALSE}
(default), raster point counts are only loaded into the \code{point_counts}
table.}

\item{union.batch}{If \code{union.rasters = TRUE}, this defines how many
rasters to load before merging data into the \code{point_counts_union}
table. The default, and minimum allowable, value is 1.}
}
\description{
First, the LAS tile is read from disk and the function checks that it is
located in one of the map zones supported by the database (each zone's
data is stored in a separate schema). Next, point heights are normalized,
any overlap between flight lines is removed, and the counts of vegetation,
ground and water points are rasterized as vertical layers defined in
\code{CERMBlidar::CERMBstrata}. Finally, the point count data are imported
into the database along with meta-data for tile extent, point density, etc.
}
