% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_load_stratum_counts}
\alias{db_load_stratum_counts}
\title{Rasterize and import points counts}
\usage{
db_load_stratum_counts(dbsettings, las, metadata.id)
}
\description{
Rasterize and import points counts
}
