% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_load_tile_metadata}
\alias{db_load_tile_metadata}
\title{Import LAS tile metadata into the database}
\usage{
db_load_tile_metadata(las, filename, dbname, tablename,
  username = "postgres")
}
\arguments{
\item{las}{A LAS object.}

\item{filename}{Path or filename from which the LAS object was read.}

\item{dbname}{The name of an existing PostgreSQL / Postgis database
(e.g. \code{'cermb_lidar'}).}

\item{tablename}{Name of the raster table in the form \code{'schema.tablename'}.}

\item{username}{Name of the user (default: 'postgres').}
}
\description{
Import LAS tile metadata into the database
}
