% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_summary}
\alias{db_summary}
\title{Summarize number of LAS tiles by map name}
\usage{
db_summary(dbsettings, with.purpose = c("general", "postfire", "all"))
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{with.purpose}{Defines the subset of data to summarize according to
image purpose. One of \code{'general'} (default) for general purpose
imagery; \code{'postfire'} for specially flown, post-fire imagery; or
\code{'all'} for all imagery. May be abbreviated.}
}
\value{
A data frame giving, for each combination of map zone (schema),
  map name and year, the number of LAS tiles imported.
}
\description{
Gets a summary of the number of LAS tiles stored in the database by
map zone (schema), map name and year.
}
