% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{pg_table_exists}
\alias{pg_table_exists}
\title{Check if a table exists in the given database}
\usage{
pg_table_exists(dbsettings, tablename)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{\link{connect_to_database}}.}

\item{tablename}{Name of the table in the form \code{schema.tablename}.}
}
\value{
\code{TRUE} if the table was found; \code{FALSE} otherwise.
}
\description{
Check if a table exists in the given database
}
\examples{
\dontrun{
dbsettings <- connect_to_database(...)
pg_table_exists(dbsettings, "postgis.bega")
}

}
