% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{pg_table_exists}
\alias{pg_table_exists}
\title{Check if a table exists in the given database}
\usage{
pg_table_exists(dbsettings, tablename)
}
\arguments{
\item{dbsettings}{A named list of database connection settings returned
by \code{db_connect_postgis} or \code{db_create_postgis}.}

\item{tablename}{Name of the table in the form \code{schema.tablename}.}
}
\value{
\code{TRUE} if the table was found; \code{FALSE} otherwise.
}
\description{
Check if a table exists in the given database
}
\examples{
\dontrun{
dbsettings <- db_connect_postgis(...)
pg_table_exists(dbsettings, "lidar.point_counts_union")
}

}
\seealso{
\code{\link{db_connect_postgis}} \code{\link{db_create_postgis}}
}
