% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{pg_table_exists}
\alias{pg_table_exists}
\title{Check if a table exists in the given database}
\usage{
pg_table_exists(dbname, tablename, username = "postgres")
}
\arguments{
\item{dbname}{The name of an existing PostgreSQL / Postgis database
(e.g. \code{'cermb_lidar'}).}

\item{tablename}{Name of the table.}

\item{username}{Name of the user (default: 'postgres').}
}
\value{
\code{TRUE} if the table was found; \code{FALSE} otherwise.
}
\description{
The check is done by using the \code{psql.exe} helper program to retrieve
metadata for the table, and checking that something was found.
}
\examples{
\dontrun{
pg_table_exists("nsw_lidar", "postgis", "bega")
}

}
